Api Rest Ok

+++ src/main/java/com/algaworks/festa/controller/ApiConvidadoController.java
@@ -1,20 +1,26 @@
 package com.algaworks.festa.controller;
 
+import java.io.Console;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.ExecutionException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.algaworks.festa.model.Convidado;
 import com.algaworks.festa.repository.Convidados;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 @RestController
 @RequestMapping("/api/convidados")
@@ -39,16 +45,19 @@ public class ApiConvidadoController {
 				HttpStatus.OK);
 	}
 
-	@RequestMapping(value = "/remove/{id}", method = RequestMethod.GET)
-	public Optional<Convidado> removeConvidado(@PathVariable("id") Long id) {
+	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+	public ResponseEntity<?> removeConvidado(@PathVariable("id") Long id) {
 		Optional<Convidado> c = convidados.findById(id);
+		if (c == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 		convidados.deleteById(id);
-		return c;
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "", method = RequestMethod.POST)
-	public  Convidado saveConvidado(Convidado c) {
-		return convidados.save(c);
+	public  ResponseEntity<?> saveConvidado(@RequestBody Convidado convidado) {
+		return new ResponseEntity<Convidado> (convidados.save(convidado), HttpStatus.OK);
 	}
 
 }
